fix(openapi): update swagger and model definitions for requestBody
