+++ src/shared/ml-thirdparty-client/models/openapi/postConsents.ts
@@ -23,14 +23,13 @@
  --------------
  ******/
 
-export enum AuthenticationType {
-  U2F = 'U2F',
-}
+import { Credential } from '../credentials'
 
-export interface AuthenticationInfo {
-  authentication: AuthenticationType
-  authenticationValue: {
-    pinValue: string
-    counter: number
-  }
+export interface RequestBody {
+  id: string
+  requestId: string
+  initiatorId: string
+  participantId: string
+  scopes: Map<string, string>
+  credential?: Credential
 }
