+++ src/shared/ml-thirdparty-client/models/openapi/putConsentRequestsById.ts
@@ -23,14 +23,11 @@
  --------------
  ******/
 
-export enum AuthenticationType {
-  U2F = 'U2F',
-}
-
-export interface AuthenticationInfo {
-  authentication: AuthenticationType
-  authenticationValue: {
-    pinValue: string
-    counter: number
-  }
+export interface RequestBody {
+  initiatorId: string
+  authChannels: string[]
+  scopes: string[]
+  callbackUri: string
+  authUri?: string
+  authToken?: string
 }
