+++ src/shared/ml-thirdparty-client/models/parties.ts
@@ -23,25 +23,179 @@
  --------------
  ******/
 
+/**
+ * The allowed values for the enumeration of party identifier type.
+ */
+export enum PartyIdType {
+  /**
+   * An MSISDN (Mobile Station International Subscriber Directory Number,
+   * that is, the phone number) is used as reference to a participant.
+   * The MSISDN identifier should be in international format according to
+   * the [ITU-T E.164 standard](https://www.itu.int/rec/T-REC-E.164/en).
+   * Optionally, the MSISDN may be prefixed by a single plus sign,
+   * indicating the international prefix.
+   */
+  MSIDN,
+
+  /**
+   * An email is used as reference to a participant.
+   * The format of the email should be according to the informational
+   * [RFC 3696](https://tools.ietf.org/html/rfc3696).
+   */
+  EMAIL,
+
+  /**
+   * A personal identifier is used as reference to a participant. Examples
+   * of personal identification are passport number, birth certificate number,
+   * and national registration number. The identifier number is added in the
+   * PartyIdentifier element. The personal identifier type is added in the
+   *  PartySubIdOrType element. to a participant.
+   */
+  PERSONAL_ID,
+
+  /**
+   * A specific Business (for example, an organization or a company) is used
+   * as reference to a participant. The BUSINESS identifier can be in any format.
+   * To make a transaction connected to a specific username or bill number in a
+   * Business, the PartySubIdOrType element should be used.
+   */
+  BUSINESS,
+
+  /**
+   * A specific device (for example, a POS or ATM) ID connected to a specific
+   * business or organization is used as reference to a Party. For referencing
+   * a specific device under a specific business or organization, use the
+   * PartySubIdOrType element.
+   */
+  DEVICE,
+
+  /**
+   * A bank account number or FSP account ID should be used as reference to
+   * a participant. The ACCOUNT_ID identifier can be in any format, as formats
+   * can greatly differ depending on country and FSP.
+   */
+  ACCOUNT_ID,
+
+  /**
+   * A bank account number or FSP account ID is used as reference to a participant.
+   * The IBAN identifier can consist of up to 34 alphanumeric characters and
+   * should be entered without whitespace.
+   */
+  IBAN,
+
+  /**
+   * An alias is used as reference to a participant. The alias should be created
+   * in the FSP as an alternative reference to an account owner. Another example
+   * of an alias is a username in the FSP system. The ALIAS identifier can be in
+   * any format. It is also possible to use the PartySubIdOrType element for
+   * identifying an account under an Alias defined by the PartyIdentifier.
+   */
+  ALIAS,
+}
+
+/**
+ * Data model for the complex type PartyComplexName.
+ */
+export interface PartyComplexName {
+  /**
+   * First name of the Party (Name Type).
+   */
+  firstName?: string
+
+  /**
+   * Middle name of the Party (Name Type).
+   */
+  middleName?: string
+
+  /**
+   * Last name of the Party (Name Type).
+   */
+  lastName?: string
+}
+
+/**
+ * Data model for the complex type PartyIdInfo.
+ */
+export interface PartyIdInfo {
+  /**
+   * Type of the identifier.
+   */
+  partyIdType: PartyIdType
+
+  /**
+   * An identifier for the Party.
+   */
+  partyIdentifier: string
+
+  /**
+   * A sub-identifier or sub-type for the Party.
+   */
+  partySubIdOrType?: string
+
+  /**
+   * FSP ID.
+   */
+  fspId?: string
+}
+
+/**
+ * Data model for the complex type PartyPersonalInfo.
+ */
+export interface PartyPersonalInfo {
+  /**
+   * First, middle and last name for the Party.
+   */
+  complexName?: PartyComplexName
+
+  /**
+   * Date of birth for the Party.
+   */
+  dateOfBirth?: string
+}
+
+/**
+ * [NEW]
+ *
+ * Data model for the complex type Account.
+ */
 export interface Account {
+  /**
+   * Address of the bank account.
+   */
   address: string
+
+  /**
+   * The currency codes defined in [ISO 4217](https://www.iso.org/iso-4217-currency-codes.html)
+   * as three-letter alphabetic codes are used as the standard naming representation for currencies.
+   *
+   * [TBC] `enum` or `string`?
+   */
   currency: string
 }
 
-export interface PartyIdInfo {
-  partyIdType: string
-  partyIdentifier: string
-  fspId: string
-}
-
+/**
+ * Data model for the complex type Party.
+ */
 export interface Party {
+  /**
+   * Party Id type, id, sub ID or type, and FSP Id.
+   */
   partyIdInfo: PartyIdInfo
-  name: string
-  accounts: Array<Account>
-}
 
-export interface PutPartiesByIdSchema {
-  partyIdType: string
-  partyIdentifier: string
-  party: Party
+  /**
+   * Used in the context of Payee Information, where the Payee happens
+   * to be a merchant accepting merchant payments.
+   */
+  merchantClassificationCode?: string
+
+  /**
+   * Display name of the Party, could be a real name or a nick name.
+   */
+  name?: string
+
+  /**
+   * Personal information used to verify identity of Party such as
+   * first, middle, last name and date of birth.
+   */
+  personalInfo?: PartyPersonalInfo
 }
